refactor(pagination): 修改类型文件名保持命名风格统一
