+++ packages/devui-vue/devui/pagination/src/pagination-types.ts
